Renamed custom types
