+++ pygame_menu/widgets/core/selection.py
@@ -33,7 +33,7 @@ __all__ = ['Selection']
 
 import pygame
 from pygame_menu.utils import assert_color
-from pygame_menu.custom_types import NumberType, ColorType, TYPE_CHECKING, Tuple2IntType, Tuple4IntType
+from pygame_menu._custom_types import NumberType, ColorType, TYPE_CHECKING, Tuple2IntType, Tuple4IntType
 
 if TYPE_CHECKING:
     from pygame_menu.widgets import Widget
